backend/requirements.txt
django==4.2.17
djangorestframework==3.15.2
psycopg2-binary
celery==5.4.0
redis==5.2.1
django-cors-headers==4.6.0
gunicorn==23.0.0
django-celery-beat==2.7.0
python-dotenv==1.0.1
dj-database-url==2.3.0
